@@ -1,78 +1,74 @@
 {
   "name": "nextjs-notion-starter-kit",
-  "version": "0.1.0",
+  "version": "2.0.0",
   "private": true,
   "description": "The perfect starter kit for building beautiful websites with Next.js and Notion.",
   "author": "Travis Fischer <travis@transitivebullsh.it>",
   "repository": "transitive-bullshit/nextjs-notion-starter-kit",
   "license": "MIT",
   "engines": {
-    "node": "16.19.0"
+    "node": ">=18"
   },
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "deploy": "vercel --prod",
-    "deps": "run-s deps:*",
-    "deps:update": "[ -z $GITHUB_ACTIONS ] && yarn add notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
-    "deps:link": "[ -z $GITHUB_ACTIONS ] && yarn link notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:link on CI'",
+    "deploy": "vercel deploy",
+    "deps:upgrade": "[ -z $GITHUB_ACTIONS ] && pnpm up -L notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
+    "deps:link": "[ -z $GITHUB_ACTIONS ] && run-s deps:link:* || echo 'Skipping deps:update on CI'",
+    "deps:unlink": "[ -z $GITHUB_ACTIONS ] && pnpm add notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
+    "deps:link:notion-types": "pnpm link ../react-notion-x/packages/notion-types",
+    "deps:link:notion-utils": "pnpm link ../react-notion-x/packages/notion-utils",
+    "deps:link:notion-client": "pnpm link ../react-notion-x/packages/notion-client",
+    "deps:link:react-notion-x": "pnpm link ../react-notion-x/packages/react-notion-x",
     "analyze": "cross-env ANALYZE=true next build",
     "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
     "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
-    "test": "run-s test:*",
+    "test": "run-p test:*",
     "test:lint": "eslint .",
-    "test:prettier": "prettier '**/*.{js,jsx,ts,tsx}' --check",
-    "posttest": "run-s build"
+    "test:prettier": "prettier '**/*.{js,jsx,ts,tsx}' --check"
   },
   "dependencies": {
-    "@google-cloud/firestore": "^4.10.1",
-    "@notionhq/client": "^0.4.13",
-    "classnames": "^2.3.1",
-    "date-fns": "^2.25.0",
-    "fathom-client": "^3.0.0",
-    "got": "^11.8.2",
-    "isomorphic-unfetch": "^3.1.0",
-    "lqip-modern": "^1.2.0",
-    "moment": "^2.29.1",
-    "next": "^11.1.2",
-    "node-fetch": "^2.6.1",
-    "notion-client": "^4.10.0",
-    "notion-to-md": "^2.2.1",
-    "notion-types": "^4.10.0",
-    "notion-utils": "^4.10.0",
-    "p-map": "^4.0.0",
-    "p-memoize": "^4.0.0",
-    "react": "^17.0.2",
+    "@fisch0920/use-dark-mode": "^2.4.0",
+    "@keyvhq/core": "^1.6.9",
+    "@keyvhq/redis": "^1.6.10",
+    "@react-icons/all-files": "^4.1.0",
+    "classnames": "^2.5.1",
+    "expiry-map": "^2.0.0",
+    "fathom-client": "^3.4.1",
+    "ky": "^1.7.2",
+    "lqip-modern": "^2.2.1",
+    "next": "^15.0.3",
+    "notion-client": "^7.1.3",
+    "notion-types": "^7.1.3",
+    "notion-utils": "^7.1.3",
+    "p-map": "^7.0.2",
+    "p-memoize": "^7.1.1",
+    "posthog-js": "^1.181.0",
+    "prismjs": "^1.29.0",
+    "react": "^18.2.0",
     "react-body-classname": "^1.3.1",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-notion-x": "^4.11.0",
-    "react-static-tweets": "^0.5.3",
-    "react-use": "^17.2.4",
-    "static-tweets": "^0.5.3",
-    "use-dark-mode": "^2.3.1"
+    "react-dom": "^18.2.0",
+    "react-notion-x": "^7.2.3",
+    "react-tweet": "^3.2.1",
+    "react-use": "^17.4.2",
+    "rss": "^1.2.2"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^11.1.2",
-    "@types/classnames": "^2.3.1",
-    "@types/node": "^16.11.2",
-    "@types/node-fetch": "^3.0.3",
-    "@types/react": "^17.0.31",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
+    "@fisch0920/eslint-config": "^1.4.0",
+    "@next/bundle-analyzer": "^15.0.2",
+    "@types/node": "^22.8.6",
+    "@types/react": "^18.0.21",
     "cross-env": "^7.0.2",
-    "eslint": "^8.31.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-config-standard-react": "^11.0.1",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-react": "^7.26.1",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.4.1",
-    "typescript": "^4.4.4"
+    "eslint": "^8.57.1",
+    "npm-run-all2": "^7.0.1",
+    "prettier": "^3.3.3",
+    "typescript": "^5.6.3"
+  },
+  "overrides": {
+    "cacheable-request": {
+      "keyv": "npm:@keyvhq/core@~1.6.6"
+    }
   }
 }
