@@ -1,22 +1,12 @@
-import React from 'react'
-import Head from 'next/head'
 import { PageHead } from './PageHead'
-
 import styles from './styles.module.css'
 
-export const ErrorPage: React.FC<{ statusCode: number }> = ({ statusCode }) => {
+export function ErrorPage({ statusCode }: { statusCode: number }) {
   const title = 'Error'
 
   return (
     <>
-      <PageHead />
-
-      <Head>
-        <meta property='og:site_name' content={title} />
-        <meta property='og:title' content={title} />
-
-        <title>{title}</title>
-      </Head>
+      <PageHead title={title} />
 
       <div className={styles.container}>
         <main className={styles.main}>
