@@ -1,8 +1,6 @@
-import React from 'react'
-
 import styles from './styles.module.css'
 
-export const GitHubShareButton: React.FC = () => {
+export function GitHubShareButton() {
   return (
     // <a
     //   href='https://github.com/transitive-bullshit/nextjs-notion-starter-kit'
