@@ -1,8 +1,8 @@
-import * as React from 'react'
 import cs from 'classnames'
+
 import styles from './styles.module.css'
 
-export const LoadingIcon = (props) => {
+export function LoadingIcon(props: any) {
   const { className, ...rest } = props
   return (
     <svg
