@@ -1,23 +1,14 @@
-import Head from 'next/head'
-import * as React from 'react'
-import * as types from 'lib/types'
-import { PageHead } from './PageHead'
+import type * as types from '@/lib/types'
 
+import { PageHead } from './PageHead'
 import styles from './styles.module.css'
 
-export const Page404: React.FC<types.PageProps> = ({ site, pageId, error }) => {
+export function Page404({ site, pageId, error }: types.PageProps) {
   const title = site?.name || 'Notion Page Not Found'
 
   return (
     <>
-      <PageHead site={site} />
-
-      <Head>
-        <meta property='og:site_name' content={title} />
-        <meta property='og:title' content={title} />
-
-        <title>{title}</title>
-      </Head>
+      <PageHead site={site} title={title} />
 
       <div className={styles.container}>
         <main className={styles.main}>
@@ -28,7 +19,8 @@ export const Page404: React.FC<types.PageProps> = ({ site, pageId, error }) => {
           ) : (
             pageId && (
               <p>
-                Make sure that Notion page "{pageId}" is publicly accessible.
+                Make sure that Notion page &quot;{pageId}&quot; is publicly
+                accessible.
               </p>
             )
           )}
