@@ -1,6 +1,5 @@
-import React from 'react'
-import Document, { Html, Head, Main, NextScript } from 'next/document'
-import { IconContext } from 'react-icons'
+import { IconContext } from '@react-icons/all-files'
+import Document, { Head, Html, Main, NextScript } from 'next/document'
 
 export default class MyDocument extends Document {
   render() {
@@ -8,38 +7,58 @@ export default class MyDocument extends Document {
       <IconContext.Provider value={{ style: { verticalAlign: 'middle' } }}>
         <Html lang='en'>
           <Head>
-            <link rel='shortcut icon' href='/favicon.png' />
-
-            <link
-              rel='apple-touch-icon'
-              sizes='180x180'
-              href='/apple-touch-icon.png'
-            />
-            <link
-              rel='icon'
-              type='image/png'
-              sizes='96x96'
-              href='/favicon-96x96.png'
-            />
+            <link rel='shortcut icon' href='/favicon.ico' />
             <link
               rel='icon'
               type='image/png'
               sizes='32x32'
-              href='/favicon-32x32.png'
-            />
-            <link
-              rel='icon'
-              type='image/png'
-              sizes='16x16'
-              href='/favicon-16x16.png'
+              href='favicon.png'
             />
 
             <link rel='manifest' href='/manifest.json' />
           </Head>
 
           <body>
-            <script src='noflash.js' />
-
+            <script
+              dangerouslySetInnerHTML={{
+                __html: `
+/** Inlined version of noflash.js from use-dark-mode */
+;(function () {
+  var storageKey = 'darkMode'
+  var classNameDark = 'dark-mode'
+  var classNameLight = 'light-mode'
+  function setClassOnDocumentBody(darkMode) {
+    document.body.classList.add(darkMode ? classNameDark : classNameLight)
+    document.body.classList.remove(darkMode ? classNameLight : classNameDark)
+  }
+  var preferDarkQuery = '(prefers-color-scheme: dark)'
+  var mql = window.matchMedia(preferDarkQuery)
+  var supportsColorSchemeQuery = mql.media === preferDarkQuery
+  var localStorageTheme = null
+  try {
+    localStorageTheme = localStorage.getItem(storageKey)
+  } catch (err) {}
+  var localStorageExists = localStorageTheme !== null
+  if (localStorageExists) {
+    localStorageTheme = JSON.parse(localStorageTheme)
+  }
+  // Determine the source of truth
+  if (localStorageExists) {
+    // source of truth from localStorage
+    setClassOnDocumentBody(localStorageTheme)
+  } else if (supportsColorSchemeQuery) {
+    // source of truth from system
+    setClassOnDocumentBody(mql.matches)
+    localStorage.setItem(storageKey, mql.matches)
+  } else {
+    // source of truth from document.body
+    var isDarkMode = document.body.classList.contains(classNameDark)
+    localStorage.setItem(storageKey, JSON.stringify(isDarkMode))
+  }
+})();
+`
+              }}
+            />
             <Main />
 
             <NextScript />
